Resolve public fuzzy service incident
According to my logs, the public fuzzy service is responding again since
the 25th of October.  The IP address of the fuzzy services are from a
range that belong to IONOS, so I assumed the services has been migrated
there and the incident is now resolved.  As such, the banners indicating
degraded services can be removed, and the blog post can be updated to
indicate that the service is now available again.